+++ src/main/java/com/telepathicgrunt/the_bumblezone/items/CrystalCannon.java
@@ -2,6 +2,8 @@ package com.telepathicgrunt.the_bumblezone.items;
 
 import com.mojang.math.Quaternion;
 import com.mojang.math.Vector3f;
+import com.telepathicgrunt.the_bumblezone.modcompat.ModChecker;
+import com.telepathicgrunt.the_bumblezone.modcompat.ProjectileDamageAttributeCompat;
 import com.telepathicgrunt.the_bumblezone.modinit.BzCriterias;
 import com.telepathicgrunt.the_bumblezone.modinit.BzItems;
 import com.telepathicgrunt.the_bumblezone.modinit.BzSounds;
@@ -58,10 +60,13 @@ public class CrystalCannon extends ProjectileWeaponItem implements Vanishable {
                     }
 
                     AbstractArrow newCrystal = BzItems.HONEY_CRYSTAL_SHARDS.createArrow(level, crystalCannon, livingEntity);
+
+                    double weaponDamage = newCrystal.getBaseDamage();
                     int power = EnchantmentHelper.getItemEnchantmentLevel(Enchantments.POWER_ARROWS, crystalCannon);
                     if (power > 0) {
-                        newCrystal.setBaseDamage(newCrystal.getBaseDamage() + (double)power * 0.5D + 0.5D);
+                        weaponDamage += (power * 0.5D) + 0.5D;
                     }
+                    newCrystal.setBaseDamage(weaponDamage);
 
                     int punch = EnchantmentHelper.getItemEnchantmentLevel(Enchantments.PUNCH_ARROWS, crystalCannon);
                     if (punch > 0) {
@@ -89,11 +94,13 @@ public class CrystalCannon extends ProjectileWeaponItem implements Vanishable {
                     Quaternion quaternion1 = new Quaternion(new Vector3f(upVector), offset, true);
                     Vector3f shootVector = new Vector3f(viewVector);
                     shootVector.transform(quaternion1);
+
+                    float weaponProjectileSpeed = 1.9F;
                     newCrystal.shoot(
                             shootVector.x(),
                             shootVector.y() + (livingEntity.getRandom().nextFloat() * 0.2f + 0.01f),
                             shootVector.z(),
-                            1.9f,
+                            weaponProjectileSpeed,
                             1);
                     level.addFreshEntity(newCrystal);
 
@@ -106,6 +113,9 @@ public class CrystalCannon extends ProjectileWeaponItem implements Vanishable {
                         BzCriterias.CRYSTAL_CANNON_FULL_TRIGGER.trigger(serverPlayer);
                     }
                 }
+
+                // Consume one extra durability
+                mutableCrystalCannon.hurtAndBreak(1, player, playerEntity -> playerEntity.broadcastBreakEvent(EquipmentSlot.MAINHAND));
             }
         }
     }